Update CIFAR10 README.md to address the issue in loading old checkpoints
Added note on how to load old CIFAR10 weights using the refactored model. 

Basically the 'module.' prefix pertaining to DataParallel wrapper needs to be removed (or the model needs to be wrapped in DataParallel), and more importantly, the extra dropout layers need to be removed except for the ones after maxpooling layers. 
These are explained in the